Refactor: CouponIssue 발급 카운트 시그니처 정리
- GlobalException.status에 INTERNAL_SERVER_ERROR 기본값을 부여해 호출부에서 message만으로도 던질 수 있게 함
- IssueCountResponse.count, CouponIssueService.countIssues 반환 타입을 Coupon.issuedQuantity(Int)에 맞춰 Long → Int로 통일

+++ high-traffic/src/main/kotlin/com/hsryuuu/hightraffic/coupon/service/CouponIssueService.kt
@@ -8,5 +8,5 @@ interface CouponIssueService {
 
     fun getCouponStatus(couponId: Long): CouponStatusResponse
 
-    fun countIssues(couponId: Long): Long
+    fun countIssues(couponId: Long): Int
 }
